Add files via upload
I am trying to create a tornado script for fivem servers, I noticed that there isnt any that I can find. So I decided to as my first project
to try and create one. If some one could help me out with the coding that be great, I like it to be able to be standalone script, able to set the time only by admins etc. Also would love it if it could be able to lift objects, I am not the best coder in the world, I just started to learn. So please be nice to me lol. Also if ur going to use this please give me credit where credit is due! Thank you. measlystorm
#2274 = Discord.